added rudimentary bar counter, doesn't update in real time yet
