+++ src/metronome.rs
@@ -20,10 +20,20 @@ pub struct MetronomeSettings {
     pub ms_delay: Arc<AtomicU64>,
     pub ts_note: Arc<AtomicU64>,
     pub ts_value: Arc<AtomicU64>,
-    pub bar_count: Arc<AtomicU64>,
     pub volume: Arc<AtomicF64>,
     pub is_running: Arc<AtomicBool>,
+    pub bar_count: Arc<AtomicU64>,
+    pub current_beat_count: Arc<AtomicU64>,
     pub error: Arc<AtomicBool>,
+}
+
+pub struct InitMetronomeSettings {
+    pub bpm: u64,
+    pub ms_delay: u64,
+    pub ts_note: u64,
+    pub ts_value: u64,
+    pub volume: f64,
+    pub is_running: bool,
 }
 
 impl Metronome {
@@ -34,9 +44,10 @@ impl Metronome {
                 ms_delay: Arc::clone(&new_settings.ms_delay),
                 ts_note: Arc::clone(&new_settings.ts_note),
                 ts_value: Arc::clone(&new_settings.ts_value),
-                bar_count: Arc::clone(&new_settings.bar_count),
                 volume: Arc::clone(&new_settings.volume),
                 is_running: Arc::clone(&new_settings.is_running),
+                bar_count: Arc::clone(&new_settings.bar_count),
+                current_beat_count: Arc::clone(&new_settings.current_beat_count),
                 error: Arc::clone(&new_settings.error),
             },
         }
@@ -66,6 +77,19 @@ impl Metronome {
                         .amplify((self.settings.volume.load(Ordering::Relaxed) / 100.0) as f32)
                         .convert_samples(),
                 );
+
+                // Bar count
+                let mut current_beat_count = self.settings.current_beat_count.load(Ordering::Relaxed);
+                if current_beat_count + 1 == self.settings.ts_note.load(Ordering::Relaxed) {
+                    self.settings.current_beat_count.swap(0, Ordering::Relaxed);
+                    let new_bar_count = self.settings.bar_count.load(Ordering::Relaxed) + 1;
+                    self.settings.bar_count.swap(new_bar_count, Ordering::Relaxed);
+                } else {
+                    current_beat_count += 1;
+                    self.settings.current_beat_count.swap(current_beat_count, Ordering::Relaxed);
+                }
+
+                // Wait
                 spin_sleep::sleep(time::Duration::from_millis(
                     self.settings.ms_delay.load(Ordering::Relaxed),
                 ));
